refactor: reservation 관련 로직을 단일 파일로 통합하여 구조 간소화
